moved animation functions to separate file
