+++ GameLib/src/component_systems/render_system.ts
@@ -49,14 +49,19 @@ class Render_System extends System_Base{
                 const {spriteSheetName} = t.getComponent('Sprite')
                 const {size} = t.getComponent('Size');
                 const {entityType} = t.getComponent('Entity');
-                const {map} = t.getComponent('Animate');
-                const {state} = t.getComponent('State');
+                
                 const ss = this.spriteSheets.resolve(spriteSheetName);
 
+                
+
                 let anim = parseInt(entityType);
-                if(map && state){
+
+                if(t.hasComponent('Animate', 'State')){
+                    const {map} = t.getComponent('Animate');
+                    const {state} = t.getComponent('State');
                     anim = map[state];
                 }
+
 
                 if(t.hasComponent('Player')){
                     //@TODO implement player logic?
@@ -65,8 +70,6 @@ class Render_System extends System_Base{
                 
                 const {sprite, imageName} = (ss.resolveSpriteData(anim, totalTime));
  
-
-                
                 const image = images[imageName];
                 if(image && sprite){
                     
@@ -87,13 +90,10 @@ class Render_System extends System_Base{
                 this.renderer.drawRect(pos.x - camera.offSets.x ,  pos.y - camera.offSets.y , size.x, size.y, 'red', false, true)
             } else if (t.hasComponent('Polygon')){
                 const {vertices} = t.getComponent('Polygon');
-                this.renderer.drawPolygon('red', false, true, camera.offSets.x ,  camera.offSets.y , vertices);
+                this.renderer.drawPolygon('red', false, true, camera.offSets.x , camera.offSets.y , vertices);
             }
             
-        })
-
-
-        
+        })        
     }
 }
